move tests to test/unit
