+++ package.json
@@ -13,7 +13,7 @@
     "protoc-gen-typescript_gapic": "build/src/cli.js"
   },
   "scripts": {
-    "test": "c8 --reporter=lcov mocha build/test",
+    "test": "c8 --reporter=lcov mocha build/test/unit",
     "codecov": "c8 --reporter=lcov mocha build/test && c8 report",
     "lint": "gts check",
     "clean": "gts clean",
@@ -37,17 +37,17 @@
     "@types/command-line-args": "^5.0.0",
     "@types/get-stdin": "^5.0.1",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^11.13.20",
+    "@types/node": "^11.13.22",
     "@types/nunjucks": "^3.1.0",
     "@types/rimraf": "^2.0.2",
     "assert-rejects": "^1.0.0",
     "c8": "^5.0.4",
-    "codecov": "^3.0.4",
+    "codecov": "^3.6.1",
     "espower-typescript": "^9.0.0",
     "google-gax": "^1.6.3",
     "gts": "^1.0.0",
     "intelli-espower-loader": "^1.0.1",
-    "mocha": "^6.0.0",
+    "mocha": "^6.2.1",
     "power-assert": "^1.6.0",
     "rimraf": "^3.0.0",
     "typescript": "~3.6.0"
