+++ typescript/test/unit/multi_proto.ts
