+++ typescript/test/unit/proto.ts
@@ -14,8 +14,8 @@
 
 import * as assert from 'assert';
 
-import * as plugin from '../../pbjs-genfiles/plugin';
-import { Proto } from '../src/schema/proto';
+import * as plugin from '../../../pbjs-genfiles/plugin';
+import { Proto } from '../../src/schema/proto';
 
 describe('schema/proto.ts', () => {
   // TODO: test service augmentation
